switch compilerOptions.jsx from react-jsx to react and move Button to src/
