+++ .gitignore
@@ -32,3 +32,5 @@ yarn-error.log*
 .env.production.local
 
 storybook-static
+
+@chromaui
