+++ src/icon/iconPaths.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export const icons = {
   user: (
     <>
